Solution v 1.0
