+++ src/components/TodoModal/TodoModal.tsx
@@ -1,28 +1,40 @@
-import React from 'react';
+import React, {useState, useEffect} from 'react';
 import { Loader } from '../Loader';
 import { User } from '../../types/User';
 import { Todo } from '../../types/Todo';
+import { getUser } from '../../api';
+
 
 type Props = {
-  activeUser: User | null;
-  activeTodo: Todo | null;
-  isModalLoading: boolean;
-  closeModal: () => void;
+  activeTodo: Todo;
+  setActiveTodo: (todo: null) => void;
 };
 
 export const TodoModal: React.FC<Props> = ({
-  activeUser,
-  isModalLoading,
   activeTodo,
-  closeModal,
+  setActiveTodo,
 }) => {
+  const [isModalLoading, setIsModalLoading] = useState(false);
+  const [activeUser, setActiveUser] = useState<User | null>(null);
+
+  const closeModal = () => {
+    setActiveTodo(null);
+  };
+
+  useEffect(() => {
+    getUser(activeTodo.userId)
+      .then(setActiveUser)
+      .finally(() => { setIsModalLoading(false) })
+  }, [activeTodo.userId]);
+
   return (
     <div className="modal is-active" data-cy="modal">
       <div className="modal-background" />
 
-      {isModalLoading
-        ? (<Loader />)
-        : (<div className="modal-card">
+      {isModalLoading ? (
+        <Loader />
+      ) : (
+        <div className="modal-card">
           <header className="modal-card-head">
             <div
               className="modal-card-title has-text-weight-medium"
@@ -46,8 +58,10 @@ export const TodoModal: React.FC<Props> = ({
             </p>
 
             <p className="block" data-cy="modal-user">
-              {/* <strong className="has-text-success">Done</strong> */}
-              <strong className="has-text-danger">Planned</strong>
+              {activeTodo.completed
+                ? (<strong className="has-text-success">Done</strong>)
+                : (<strong className="has-text-danger">Planned</strong>)
+              }
 
               {' by '}
 
@@ -55,8 +69,7 @@ export const TodoModal: React.FC<Props> = ({
             </p>
           </div>
         </div>
-        )
-      }
+      )}
     </div>
   );
 };
